Move stats entities to correct package
