+++ src/main/java/space/npstr/wolfia/commands/util/ReplayCommand.java
@@ -20,12 +20,12 @@ package space.npstr.wolfia.commands.util;
 import space.npstr.wolfia.commands.BaseCommand;
 import space.npstr.wolfia.commands.CommandContext;
 import space.npstr.wolfia.commands.PublicCommand;
-import space.npstr.wolfia.db.entities.stats.ActionStats;
-import space.npstr.wolfia.db.entities.stats.GameStats;
-import space.npstr.wolfia.db.entities.stats.TeamStats;
 import space.npstr.wolfia.domain.Command;
+import space.npstr.wolfia.domain.stats.ActionStats;
+import space.npstr.wolfia.domain.stats.GameStats;
 import space.npstr.wolfia.domain.stats.StatsProvider;
 import space.npstr.wolfia.domain.stats.StatsRender;
+import space.npstr.wolfia.domain.stats.TeamStats;
 import space.npstr.wolfia.game.definitions.Games;
 import space.npstr.wolfia.game.tools.NiceEmbedBuilder;
 import space.npstr.wolfia.utils.discord.Emojis;
