+++ src/main/java/space/npstr/wolfia/domain/stats/StatsRender.java
@@ -26,10 +26,6 @@ import org.springframework.stereotype.Component;
 import space.npstr.sqlsauce.entities.discord.DiscordUser;
 import space.npstr.wolfia.commands.Context;
 import space.npstr.wolfia.commands.MessageContext;
-import space.npstr.wolfia.db.entities.stats.ActionStats;
-import space.npstr.wolfia.db.entities.stats.GameStats;
-import space.npstr.wolfia.db.entities.stats.PlayerStats;
-import space.npstr.wolfia.db.entities.stats.TeamStats;
 import space.npstr.wolfia.domain.UserCache;
 import space.npstr.wolfia.domain.settings.GuildSettings;
 import space.npstr.wolfia.domain.settings.GuildSettingsService;
