+++ src/test/java/space/npstr/wolfia/domain/stats/StatsRepositoryTest.java
@@ -20,10 +20,6 @@ package space.npstr.wolfia.domain.stats;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import space.npstr.wolfia.ApplicationTest;
-import space.npstr.wolfia.db.entities.stats.ActionStats;
-import space.npstr.wolfia.db.entities.stats.GameStats;
-import space.npstr.wolfia.db.entities.stats.PlayerStats;
-import space.npstr.wolfia.db.entities.stats.TeamStats;
 import space.npstr.wolfia.game.GameInfo;
 import space.npstr.wolfia.game.definitions.Actions;
 import space.npstr.wolfia.game.definitions.Alignments;
